Build record repo and some other stuff

+++ app/src/main/java/org/cooltey/punchbabycounter/database/Record.kt
@@ -7,8 +7,8 @@ import java.util.*
 
 @Entity
 data class Record(
-    @PrimaryKey(autoGenerate = true) val uid: Int = 0,
-    @ColumnInfo(name = "user_id") val userId: Int,
+    @PrimaryKey(autoGenerate = true) val uid: Long = 0,
+    @ColumnInfo(name = "user_id") val userId: Long,
     @ColumnInfo(name = "level_1") val level1: Long, // Level 1 punch
     @ColumnInfo(name = "level_2") val level2: Long, // Level 2 punch
     @ColumnInfo(name = "date") val date: Date,
